+++ handlers/writings/cancelTask.go
@@ -1,0 +1,29 @@
+package writings
+
+import (
+	"fmt"
+	"net/http"
+	"strconv"
+
+	"github.com/arran4/goa4web/handlers"
+	"github.com/arran4/goa4web/internal/tasks"
+	"github.com/gorilla/mux"
+)
+
+// CancelTask cancels comment editing.
+type CancelTask struct{ tasks.TaskString }
+
+var cancelTask = &CancelTask{TaskString: TaskCancel}
+
+// CancelTask is only used to abort editing, implementing tasks.Task ensures it
+// fits the routing interface even though no additional behaviour is required.
+var _ tasks.Task = (*CancelTask)(nil)
+
+func (CancelTask) Action(w http.ResponseWriter, r *http.Request) any {
+	vars := mux.Vars(r)
+	articleId, err := strconv.Atoi(vars["article"])
+	if err != nil {
+		return fmt.Errorf("article id parse %w", handlers.ErrRedirectOnSamePageHandler(err))
+	}
+	return handlers.RedirectHandler(fmt.Sprintf("/writings/article/%d", articleId))
+}
