+++ handlers/writings/userAllowTask.go
@@ -1,0 +1,64 @@
+package writings
+
+import (
+	"database/sql"
+	"fmt"
+	"net/http"
+
+	"github.com/arran4/goa4web/core/common"
+	"github.com/arran4/goa4web/core/consts"
+	"github.com/arran4/goa4web/internal/db"
+	"github.com/arran4/goa4web/internal/eventbus"
+	notif "github.com/arran4/goa4web/internal/notifications"
+	"github.com/arran4/goa4web/internal/tasks"
+)
+
+// UserAllowTask grants a user a permission.
+type UserAllowTask struct{ tasks.TaskString }
+
+var userAllowTask = &UserAllowTask{TaskString: TaskUserAllow}
+
+var _ tasks.Task = (*UserAllowTask)(nil)
+var _ notif.TargetUsersNotificationProvider = (*UserAllowTask)(nil)
+
+func (UserAllowTask) Action(w http.ResponseWriter, r *http.Request) any {
+	if r.URL.Path == "/admin/writings/users/roles" {
+		queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
+		username := r.PostFormValue("username")
+		role := r.PostFormValue("role")
+		u, err := queries.GetUserByUsername(r.Context(), sql.NullString{Valid: true, String: username})
+		if err != nil {
+			return fmt.Errorf("GetUserByUsername: %w", err)
+		}
+
+		if err := queries.CreateUserRole(r.Context(), db.CreateUserRoleParams{
+			UsersIdusers: u.Idusers,
+			Name:         role,
+		}); err != nil {
+			return fmt.Errorf("permissionUserAllow: %w", err)
+		}
+		return nil
+	}
+	// TODO: inline UsersPermissionsPermissionUserAllowPage or provide a
+	// dedicated task if possible.
+	return http.HandlerFunc(UsersPermissionsPermissionUserAllowPage)
+}
+
+func (UserAllowTask) TargetUserIDs(evt eventbus.TaskEvent) ([]int32, error) {
+	if id, ok := evt.Data["targetUserID"].(int32); ok {
+		return []int32{id}, nil
+	}
+	if id, ok := evt.Data["targetUserID"].(int); ok {
+		return []int32{int32(id)}, nil
+	}
+	return nil, fmt.Errorf("target user id not provided")
+}
+
+func (UserAllowTask) TargetEmailTemplate() *notif.EmailTemplates {
+	return notif.NewEmailTemplates("setUserRoleEmail")
+}
+
+func (UserAllowTask) TargetInternalNotificationTemplate() *string {
+	v := notif.NotificationTemplateFilenameGenerator("set_user_role")
+	return &v
+}
