+++ handlers/writings/writingsAdminUserLevelsPage.go
@@ -6,7 +6,6 @@ import (
 	"github.com/arran4/goa4web/core/consts"
 	"log"
 	"net/http"
-	"strconv"
 
 	"github.com/arran4/goa4web/core/common"
 	"github.com/arran4/goa4web/handlers"
@@ -41,43 +40,4 @@ func AdminUserRolesPage(w http.ResponseWriter, r *http.Request) {
 	data.UserLevels = rows
 
 	handlers.TemplateHandler(w, r, "adminUserRolesPage.gohtml", data)
-}
-
-func AdminUserLevelsAllowActionPage(w http.ResponseWriter, r *http.Request) {
-	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
-	username := r.PostFormValue("username")
-	role := r.PostFormValue("role")
-	u, err := queries.GetUserByUsername(r.Context(), sql.NullString{Valid: true, String: username})
-	if err != nil {
-		log.Printf("GetUserByUsername Error: %s", err)
-		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
-		return
-	}
-
-	if err := queries.CreateUserRole(r.Context(), db.CreateUserRoleParams{
-		UsersIdusers: u.Idusers,
-		Name:         role,
-	}); err != nil {
-		log.Printf("permissionUserAllow Error: %s", err)
-		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
-		return
-	}
-	handlers.TaskDoneAutoRefreshPage(w, r)
-}
-
-func AdminUserLevelsRemoveActionPage(w http.ResponseWriter, r *http.Request) {
-	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
-	permid := r.PostFormValue("permid")
-	permidi, err := strconv.Atoi(permid)
-	if err != nil {
-		log.Printf("strconv.Atoi(permid Error: %s", err)
-		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
-		return
-	}
-	if err := queries.DeleteUserRole(r.Context(), int32(permidi)); err != nil {
-		log.Printf("permissionUserDisallow Error: %s", err)
-		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
-		return
-	}
-	handlers.TaskDoneAutoRefreshPage(w, r)
 }
