+++ handlers/writings/writingsArticleEditPage.go
@@ -1,22 +1,14 @@
 package writings
 
 import (
-	"database/sql"
-	"fmt"
 	"github.com/arran4/goa4web/core/consts"
-	"log"
 	"net/http"
-	"strconv"
-
-	"github.com/arran4/goa4web/core/common"
-	"github.com/arran4/goa4web/handlers"
-	"github.com/arran4/goa4web/internal/db"
-	notif "github.com/arran4/goa4web/internal/notifications"
-	"github.com/arran4/goa4web/workers/searchworker"
-	"strings"
 
 	"github.com/arran4/goa4web/config"
 	"github.com/arran4/goa4web/core"
+	"github.com/arran4/goa4web/core/common"
+	"github.com/arran4/goa4web/handlers"
+	"github.com/arran4/goa4web/internal/db"
 )
 
 func ArticleEditPage(w http.ResponseWriter, r *http.Request) {
@@ -55,63 +47,4 @@ func ArticleEditPage(w http.ResponseWriter, r *http.Request) {
 	data.Languages = languageRows
 
 	handlers.TemplateHandler(w, r, "articleEditPage.gohtml", data)
-}
-
-func ArticleEditActionPage(w http.ResponseWriter, r *http.Request) {
-	// RequireWritingAuthor middleware loads the writing and validates access.
-	writing := r.Context().Value(consts.KeyWriting).(*db.GetWritingByIdForUserDescendingByPublishedDateRow)
-
-	languageId, _ := strconv.Atoi(r.PostFormValue("language"))
-	private, _ := strconv.ParseBool(r.PostFormValue("isitprivate"))
-	title := r.PostFormValue("title")
-	abstract := r.PostFormValue("abstract")
-	body := r.PostFormValue("body")
-
-	cd := r.Context().Value(consts.KeyCoreData).(*common.CoreData)
-	queries := cd.Queries()
-
-	if err := queries.UpdateWriting(r.Context(), db.UpdateWritingParams{
-		Title:              sql.NullString{Valid: true, String: title},
-		Abstract:           sql.NullString{Valid: true, String: abstract},
-		Writing:            sql.NullString{Valid: true, String: body},
-		Private:            sql.NullBool{Valid: true, Bool: private},
-		LanguageIdlanguage: int32(languageId),
-		Idwriting:          writing.Idwriting,
-	}); err != nil {
-		log.Printf("updateWriting Error: %s", err)
-		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
-		return
-	}
-
-	if cd, ok := r.Context().Value(consts.KeyCoreData).(*common.CoreData); ok {
-		if evt := cd.Event(); evt != nil {
-			if evt.Data == nil {
-				evt.Data = map[string]any{}
-			}
-			author := ""
-			if writing.Writerusername.Valid {
-				author = writing.Writerusername.String
-			}
-			evt.Data["Title"] = title
-			evt.Data["Author"] = author
-			evt.Data["PostURL"] = cd.AbsoluteURL(fmt.Sprintf("/writings/article/%d", writing.Idwriting))
-			evt.Data["target"] = notif.Target{Type: "writing", ID: writing.Idwriting}
-		}
-	}
-
-	if err := queries.WritingSearchDelete(r.Context(), writing.Idwriting); err != nil {
-		log.Printf("writingSearchDelete Error: %s", err)
-		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
-		return
-	}
-
-	fullText := strings.Join([]string{abstract, title, body}, " ")
-	if cd, ok := r.Context().Value(consts.KeyCoreData).(*common.CoreData); ok {
-		if evt := cd.Event(); evt != nil {
-			if evt.Data == nil {
-				evt.Data = map[string]any{}
-			}
-			evt.Data[searchworker.EventKey] = searchworker.IndexEventData{Type: searchworker.TypeWriting, ID: writing.Idwriting, Text: fullText}
-		}
-	}
 }
